requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
Django==1.11.13
django-cors-headers==2.2.0
djangorestframework==3.8.2
idna==2.7
pytz==2018.4
requests==2.19.1
urllib3==1.23
